+++ customer/customer_list.php
@@ -33,7 +33,7 @@
                             <td class ='category-options'>". $row["apellido1"]." ". $row["apellido2"] ."</td>
                             <td><a class='category-options details' href='customer_detail.php?dni=".$row["dni"]."'><i class='bx bx-info-circle' ></i></a></td>
                             <td><a class='category-options edit' href='customer_list_update.php?dni=".$row["dni"]."'><i class='bx bx-edit-alt' ></i></a></td>
-                            <td><a class='category-options delete' href='customer_list.php?deleteid=".$row["dni"]."'><i class='bx bx-x' ></i></a></td>
+                            <td><a class='category-options delete' onClick=\"return confirm('Are you sure?')\" href='customer_list.php?deleteid=".$row["dni"]."'><i class='bx bx-x' ></i></a></td>
                         <tr>";
                 } 
             ?>
