+++ product/product_list.php
@@ -40,7 +40,7 @@
                             <td class ='category-options'>". $row["precio"] ."$</td>
                             <td><a class='category-options details' href='product_detail.php?id=".$row["id"]."'><i class='bx bx-info-circle' ></i></a></td>
                             <td><a class='category-options edit' href='product_list_update.php?id=".$row["id"]."'><i class='bx bx-edit-alt' ></i></a></td>
-                            <td><a class='category-options delete' href='product_list.php?deleteid=".$row["id"]."'><i class='bx bx-x' ></i></a></td>
+                            <td><a class='category-options delete' onClick=\"return confirm('Are you sure?')\" href='product_list.php?deleteid=".$row["id"]."'><i class='bx bx-x' ></i></a></td>
                         <tr>";
                 } 
             ?>
